+++ .idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="..\:/AndroidStudioProjects/SWIMSTest/app/src/main/res/layout/activity_login.xml" value="0.20677083333333332" />
         <entry key="..\:/AndroidStudioProjects/SWIMSTest/app/src/main/res/layout/activity_main.xml" value="0.2364179037174423" />
+        <entry key="..\:/AndroidStudioProjects/SWIMSTest/app/src/main/res/layout/activity_sign_in.xml" value="0.20677083333333332" />
         <entry key="..\:/AndroidStudioProjects/SWIMSTest/app/src/main/res/xml/file_paths.xml" value="0.20677083333333332" />
       </map>
     </option>
